+++ UnityEngine.TestRunner/TestRunner/PlaymodeTestsControllerSettings.cs
@@ -15,6 +15,9 @@ namespace UnityEngine.TestTools.TestRunner
         public bool sceneBased;
         public string originalScene;
         public string bootstrapScene;
+        public bool runInBackgroundValue;
+        public bool consoleErrorPaused;
+
 
         public static PlaymodeTestsControllerSettings CreateRunnerSettings(RuntimeTestRunnerFilter[] filters)
         {